+++ tests-ng/ignore-empty.sh
@@ -28,30 +28,33 @@ cur=$(dirname "$(${rl} "${0}")")
 # dotdrop path can be pass as argument
 ddpath="${cur}/../"
 [ "${1}" != "" ] && ddpath="${1}"
-[ ! -d ${ddpath} ] && echo "ddpath \"${ddpath}\" is not a directory" && exit 1
+[ ! -d "${ddpath}" ] && echo "ddpath \"${ddpath}\" is not a directory" && exit 1
 
 export PYTHONPATH="${ddpath}:${PYTHONPATH}"
 bin="python3 -m dotdrop.dotdrop"
-hash coverage 2>/dev/null && bin="coverage run -a --source=dotdrop -m dotdrop.dotdrop" || true
+if hash coverage 2>/dev/null; then
+  bin="coverage run -a --source=dotdrop -m dotdrop.dotdrop"
+fi
 
 echo "dotdrop path: ${ddpath}"
 echo "pythonpath: ${PYTHONPATH}"
 
 # get the helpers
-source ${cur}/helpers
+# shellcheck source=tests-ng/helpers
+source "${cur}"/helpers
 
-echo -e "$(tput setaf 6)==> RUNNING $(basename $BASH_SOURCE) <==$(tput sgr0)"
+echo -e "$(tput setaf 6)==> RUNNING $(basename "${BASH_SOURCE[0]}") <==$(tput sgr0)"
 
 ################################################################
 # this is the test
 ################################################################
 
 # the dotfile source
-tmps=`mktemp -d --suffix='-dotdrop-tests' || mktemp -d`
-mkdir -p ${tmps}/dotfiles
+tmps=$(mktemp -d --suffix='-dotdrop-tests' || mktemp -d)
+mkdir -p "${tmps}"/dotfiles
 #echo "dotfile source: ${tmps}"
 # the dotfile destination
-tmpd=`mktemp -d --suffix='-dotdrop-tests' || mktemp -d`
+tmpd=$(mktemp -d --suffix='-dotdrop-tests' || mktemp -d)
 #echo "dotfile destination: ${tmpd}"
 
 clear_on_exit "${tmps}"
@@ -61,7 +64,7 @@ clear_on_exit "${tmpd}"
 cfg="${tmps}/config.yaml"
 
 # globally
-cat > ${cfg} << _EOF
+cat > "${cfg}" << _EOF
 config:
   backup: true
   create: true
@@ -79,19 +82,19 @@ _EOF
 #cat ${cfg}
 
 # create the dotfile
-mkdir -p ${tmps}/dotfiles/d1
-echo "{#@@ should be stripped @@#}" > ${tmps}/dotfiles/d1/empty
-echo "not empty" > ${tmps}/dotfiles/d1/notempty
+mkdir -p "${tmps}"/dotfiles/d1
+echo "{#@@ should be stripped @@#}" > "${tmps}"/dotfiles/d1/empty
+echo "not empty" > "${tmps}"/dotfiles/d1/notempty
 
 # install
-cd ${ddpath} | ${bin} install -f -c ${cfg} -p p1 -V
+cd "${ddpath}" | ${bin} install -f -c "${cfg}" -p p1 -V
 
 # test existence
-[ -e ${tmpd}/d1/empty ] && echo 'empty should not exist' && exit 1
-[ ! -e ${tmpd}/d1/notempty ] && echo 'not empty should exist' && exit 1
+[ -e "${tmpd}"/d1/empty ] && echo 'empty should not exist' && exit 1
+[ ! -e "${tmpd}"/d1/notempty ] && echo 'not empty should exist' && exit 1
 
 # through the dotfile
-cat > ${cfg} << _EOF
+cat > "${cfg}" << _EOF
 config:
   backup: true
   create: true
@@ -110,14 +113,14 @@ _EOF
 #cat ${cfg}
 
 # clean destination
-rm -rf ${tmpd}/*
+rm -rf "${tmpd:?}"/*
 
 # install
-cd ${ddpath} | ${bin} install -f -c ${cfg} -p p1 -V
+cd "${ddpath}" | ${bin} install -f -c "${cfg}" -p p1 -V
 
 # test existence
-[ -e ${tmpd}/d1/empty ] && echo 'empty should not exist' && exit 1
-[ ! -e ${tmpd}/d1/notempty ] && echo 'not empty should exist' && exit 1
+[ -e "${tmpd}"/d1/empty ] && echo 'empty should not exist' && exit 1
+[ ! -e "${tmpd}"/d1/notempty ] && echo 'not empty should exist' && exit 1
 
 echo "OK"
 exit 0
